pyproject.toml
[project]
name = "ops-agent"
version = "0.1.0"
description = "MCP-based intelligent AIOps diagnosis multi-agent system"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    # MCP
    "mcp[cli]>=1.0.0",
    # LangGraph / LangChain
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    # RAG
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    # Anomaly Detection
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    # Utils
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]
